Request accounts

+++ app/src/main/java/com/storiqa/storiqawallet/data/IUserDataStorage.kt
@@ -2,7 +2,7 @@ package com.storiqa.storiqawallet.data
 
 interface IUserDataStorage {
 
-    var id: Int
+    var id: Long
     var email: String
     var firstName: String
     var lastName: String
